requirements.txt
appdirs==1.4.0
Django==1.10.5
django-cors-headers==2.0.2
djangorestframework==3.5.4
gunicorn==19.7.1
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.1.10
six==1.10.0
virtualenv==15.1.0
